# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SoleFlipper API",
    "description": "\n# SoleFlipper API Documentation\n\n## Overview\nProfessional sneaker reselling management system with inventory tracking, \nautomated imports, and analytics integration.\n\n## Architecture\n- **FastAPI**: Modern async Python framework\n- **PostgreSQL**: Primary database with schema separation\n- **n8n Integration**: Webhook-based automation\n- **Metabase**: Analytics and reporting\n- **Domain-Driven Design**: Clean architecture patterns\n\n## Authentication\nCurrently, most endpoints are open. Authentication will be added in future versions.\n\n## Rate Limiting\nRate limiting is planned but not yet implemented.\n\n## Data Import Sources\n- **StockX**: Sales and transaction data\n- **Notion**: Inventory management database\n- **Manual**: CSV/Excel file uploads with auto-detection\n\n## Error Handling\nAll endpoints return structured error responses with:\n- `error.type`: Error category\n- `error.message`: Human-readable description\n- `error.details`: Additional context (optional)\n\n## Pagination\nList endpoints support query parameters:\n- `limit`: Number of items (default: 50, max: 1000)\n- `offset`: Number of items to skip (default: 0)\n        ",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    },
    {
      "url": "https://api.soleflip.com",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/api/v1/integration/webhooks/stockx/upload": {
      "post": {
        "tags": [
          "StockX Integration"
        ],
        "summary": "Upload Stockx File",
        "operationId": "upload_stockx_file_api_v1_integration_webhooks_stockx_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_stockx_file_api_v1_integration_webhooks_stockx_upload_post"
              },
              "example": {
                "file": "stockx_sales_report.csv",
                "batch_size": 1000,
                "validate_only": false
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                },
                "examples": {
                  "processing_started": {
                    "summary": "Import processing started",
                    "value": {
                      "status": "processing_started",
                      "message": "File uploaded successfully. Processing started in background.",
                      "batch_id": "123e4567-e89b-12d3-a456-426614174000",
                      "filename": "stockx_sales_report.csv",
                      "record_count": 150,
                      "check_status_url": "/api/v1/integration/webhooks/import-status/123e4567-e89b-12d3-a456-426614174000"
                    }
                  },
                  "validation_only": {
                    "summary": "Validation only result",
                    "value": {
                      "status": "validated",
                      "valid": true,
                      "errors": [],
                      "warnings": [
                        "Some items missing optional fields"
                      ],
                      "record_count": 150
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Body_upload_stockx_file_api_v1_integration_webhooks_stockx_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          },
          "validate_only": {
            "type": "boolean",
            "title": "Validate Only",
            "default": false
          },
          "batch_size": {
            "type": "integer",
            "title": "Batch Size",
            "default": 1000
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_stockx_file_api_v1_integration_webhooks_stockx_upload_post"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UploadResponse": {
        "properties": {
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "total_records": {
            "type": "integer",
            "title": "Total Records"
          },
          "validation_errors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Validation Errors"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "imported": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Imported"
          },
          "batch_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Batch Id"
          }
        },
        "type": "object",
        "required": [
          "filename",
          "total_records",
          "validation_errors",
          "status",
          "message"
        ],
        "title": "UploadResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}